requirements-core.txt
fastapi==0.128.7
uvicorn==0.40.0
structlog==25.5.0
pydantic==2.12.5
pydantic-settings==2.12.0
httpx==0.28.1
openai==2.18.0
python-dotenv==1.2.1
pytest==9.0.2
ruff==0.14.1
PyJWT==2.10.1
cryptography==46.0.3
